feat: dashboard grid layout with quick-entry widgets
Restructure Dashboard from stacked cards to responsive CSS grid.
Add QuickBrainCapture (thought/URL) and QuickTaskWidget (compact task form)
at top of page. Extract TaskAddForm as shared component with compact/full
modes, replacing window.prompt() with inline input.
