+++ .changelog/v0.25.x.md
@@ -14,6 +14,13 @@ Peer apps display and remote agent streaming on the Instances page — see what'
 - **Socket.IO peer relay** — local server connects as a Socket.IO client to each online peer, subscribes to agent events, and forwards them to the Instances page
 - **AgentCard remote mode** — `remote` prop disables local-only features (process stats polling, full output fetch, feedback, kill/delete buttons) and shows a peer name badge
 
+### Dashboard grid layout redesign
+- **Responsive grid layout** — Dashboard sections now use a CSS grid instead of vertically stacked full-width cards: System Health + Quick Stats side-by-side at lg, CoS/Goals/Tasks in a 3-column row, Activity Streak + Heatmap side-by-side; everything stacks single-column on mobile
+- **Quick Brain Capture widget** — capture thoughts or paste URLs directly from the dashboard; auto-detects URLs vs. text and routes to the appropriate Brain API
+- **Quick Task widget** — add CoS tasks from the dashboard via a compact form with expandable "More options" for provider, model, worktree, screenshots, etc.
+- **Extracted TaskAddForm component** — the task add form is now a shared component (`TaskAddForm.jsx`) used by both the CoS Tasks page (full mode) and the dashboard widget (compact mode)
+- **Replaced `window.prompt()`** — template save now uses an inline text input instead of `window.prompt()`, complying with the no-browser-dialogs rule
+
 ## Improvements
 
 ### Event-driven CoS evaluation architecture
