+++ client/src/pages/Dashboard.jsx
@@ -7,6 +7,8 @@ import CosDashboardWidget from '../components/CosDashboardWidget';
 import GoalProgressWidget from '../components/GoalProgressWidget';
 import UpcomingTasksWidget from '../components/UpcomingTasksWidget';
 import DecisionLogWidget from '../components/DecisionLogWidget';
+import QuickBrainCapture from '../components/QuickBrainCapture';
+import QuickTaskWidget from '../components/QuickTaskWidget';
 import * as api from '../services/api';
 import socket from '../services/socket';
 
@@ -33,7 +35,6 @@ export default function Dashboard() {
   useEffect(() => {
     fetchData();
 
-    // Listen for apps changes via WebSocket instead of polling
     const handleAppsChanged = () => {
       fetchData();
     };
@@ -50,7 +51,7 @@ export default function Dashboard() {
     [apps]
   );
 
-  // Memoize derived stats to prevent recalculation on every render (exclude archived)
+  // Memoize derived stats (exclude archived)
   const activeApps = useMemo(() => apps.filter(a => !a.archived), [apps]);
   const appStats = useMemo(() => ({
     total: activeApps.length,
@@ -68,9 +69,9 @@ export default function Dashboard() {
   }
 
   return (
-    <div>
+    <div className="space-y-6">
       {/* Header */}
-      <div className="flex flex-col sm:flex-row sm:items-center justify-between gap-2 sm:gap-4 mb-6">
+      <div className="flex flex-col sm:flex-row sm:items-center justify-between gap-2 sm:gap-4">
         <div>
           <h2 className="text-2xl font-bold text-white">Dashboard</h2>
           <p className="text-gray-500 text-sm sm:text-base">
@@ -86,37 +87,48 @@ export default function Dashboard() {
 
       {/* Error */}
       {error && (
-        <div className="mb-6 p-4 bg-port-error/20 border border-port-error rounded-lg text-port-error">
+        <div className="p-4 bg-port-error/20 border border-port-error rounded-lg text-port-error">
           {error}
         </div>
       )}
 
-      {/* System Health Widget */}
-      <div className="mb-6">
-        <SystemHealthWidget />
+      {/* Row 1 — Quick entry */}
+      <div className="grid grid-cols-1 md:grid-cols-2 gap-4">
+        <QuickBrainCapture />
+        <QuickTaskWidget />
       </div>
 
-      {/* CoS Status Widget */}
-      <div className="mb-6">
+      {/* Row 2 — Status at a glance */}
+      <div className="grid grid-cols-1 lg:grid-cols-12 gap-4">
+        <div className="lg:col-span-8">
+          <SystemHealthWidget />
+        </div>
+        {apps.length > 0 && (
+          <div className="lg:col-span-4">
+            <div className="bg-port-card border border-port-border rounded-xl p-4 h-full">
+              <h3 className="text-sm font-semibold text-white mb-3">Quick Stats</h3>
+              <div className="grid grid-cols-2 gap-2">
+                <StatCard label="Total Apps" value={appStats.total} icon="📦" />
+                <StatCard label="Online" value={appStats.online} icon="🟢" />
+                <StatCard label="Stopped" value={appStats.stopped} icon="🟡" />
+                <StatCard label="Not Started" value={appStats.notStarted} icon="⚪" />
+              </div>
+            </div>
+          </div>
+        )}
+      </div>
+
+      {/* Row 3 — AI/CoS intelligence */}
+      <div className="grid grid-cols-1 lg:grid-cols-3 gap-4">
         <CosDashboardWidget />
-      </div>
-
-      {/* Goal Progress Widget */}
-      <div className="mb-6">
         <GoalProgressWidget />
-      </div>
-
-      {/* Upcoming Tasks Widget */}
-      <div className="mb-6">
         <UpcomingTasksWidget />
       </div>
 
-      {/* Decision Log Widget */}
-      <div className="mb-6">
-        <DecisionLogWidget />
-      </div>
+      {/* Row 4 — Decision log */}
+      <DecisionLogWidget />
 
-      {/* App Grid */}
+      {/* Row 5 — Apps */}
       {apps.length === 0 ? (
         <div className="bg-port-card border border-port-border rounded-xl p-8 sm:p-12 text-center">
           <div className="text-4xl mb-4">📦</div>
@@ -139,83 +151,62 @@ export default function Dashboard() {
         </div>
       )}
 
-      {/* Activity Streak */}
-      {usage && (usage.currentStreak > 0 || usage.longestStreak > 0) && (
-        <div className="mt-6 sm:mt-8 bg-port-card border border-port-border rounded-xl p-4 sm:p-6">
-          <div className="flex flex-col sm:flex-row sm:items-center gap-3 sm:gap-4">
-            <div className="flex items-center gap-3">
-              <div className="text-3xl sm:text-4xl" aria-hidden="true">
-                {usage.currentStreak >= 7 ? '🔥' : usage.currentStreak >= 3 ? '⚡' : '✨'}
-              </div>
-              <div>
-                <div className="text-xl sm:text-2xl font-bold text-white">
-                  {usage.currentStreak} day{usage.currentStreak !== 1 ? 's' : ''}
+      {/* Row 6 — Activity data */}
+      {(usage?.currentStreak > 0 || usage?.longestStreak > 0 || usage?.hourlyActivity) && (
+        <div className="grid grid-cols-1 md:grid-cols-2 gap-4">
+          {/* Activity Streak */}
+          {usage && (usage.currentStreak > 0 || usage.longestStreak > 0) && (
+            <div className="bg-port-card border border-port-border rounded-xl p-4 sm:p-6">
+              <div className="flex flex-col sm:flex-row sm:items-center gap-3 sm:gap-4">
+                <div className="flex items-center gap-3">
+                  <div className="text-3xl sm:text-4xl" aria-hidden="true">
+                    {usage.currentStreak >= 7 ? '🔥' : usage.currentStreak >= 3 ? '⚡' : '✨'}
+                  </div>
+                  <div>
+                    <div className="text-xl sm:text-2xl font-bold text-white">
+                      {usage.currentStreak} day{usage.currentStreak !== 1 ? 's' : ''}
+                    </div>
+                    <div className="text-xs sm:text-sm text-gray-500">Current streak</div>
+                  </div>
                 </div>
-                <div className="text-xs sm:text-sm text-gray-500">Current streak</div>
+                {usage.longestStreak > usage.currentStreak && (
+                  <div className="sm:ml-auto text-left sm:text-right">
+                    <div className="text-base sm:text-lg font-semibold text-port-accent">
+                      {usage.longestStreak} days
+                    </div>
+                    <div className="text-xs text-gray-500">Longest streak</div>
+                  </div>
+                )}
+                {usage.currentStreak === usage.longestStreak && usage.currentStreak > 0 && (
+                  <div className="sm:ml-auto px-3 py-1.5 bg-port-success/20 text-port-success text-xs sm:text-sm rounded-full w-fit">
+                    Personal best!
+                  </div>
+                )}
+              </div>
+              {/* Mini streak visualization */}
+              <div className="mt-4 flex gap-1.5 sm:gap-1">
+                {usage.last7Days?.map((day) => (
+                  <div
+                    key={day.date}
+                    className={`flex-1 h-3 sm:h-2 rounded-full ${
+                      day.sessions > 0 ? 'bg-port-success' : 'bg-port-border'
+                    }`}
+                    title={`${day.label}: ${day.sessions} sessions`}
+                  />
+                ))}
+              </div>
+              <div className="mt-1 flex justify-between text-xs text-gray-500">
+                <span>7 days ago</span>
+                <span>Today</span>
               </div>
             </div>
-            {usage.longestStreak > usage.currentStreak && (
-              <div className="sm:ml-auto text-left sm:text-right">
-                <div className="text-base sm:text-lg font-semibold text-port-accent">
-                  {usage.longestStreak} days
-                </div>
-                <div className="text-xs text-gray-500">Longest streak</div>
-              </div>
-            )}
-            {usage.currentStreak === usage.longestStreak && usage.currentStreak > 0 && (
-              <div className="sm:ml-auto px-3 py-1.5 bg-port-success/20 text-port-success text-xs sm:text-sm rounded-full w-fit">
-                Personal best!
-              </div>
-            )}
-          </div>
-          {/* Mini streak visualization */}
-          <div className="mt-4 flex gap-1.5 sm:gap-1">
-            {usage.last7Days?.map((day) => (
-              <div
-                key={day.date}
-                className={`flex-1 h-3 sm:h-2 rounded-full ${
-                  day.sessions > 0 ? 'bg-port-success' : 'bg-port-border'
-                }`}
-                title={`${day.label}: ${day.sessions} sessions`}
-              />
-            ))}
-          </div>
-          <div className="mt-1 flex justify-between text-xs text-gray-500">
-            <span>7 days ago</span>
-            <span>Today</span>
-          </div>
-        </div>
-      )}
+          )}
 
-      {/* Quick Stats */}
-      {apps.length > 0 && (
-        <div className="mt-6 sm:mt-8 grid grid-cols-2 md:grid-cols-4 gap-2 sm:gap-4">
-          <StatCard
-            label="Total Apps"
-            value={appStats.total}
-            icon="📦"
-          />
-          <StatCard
-            label="Online"
-            value={appStats.online}
-            icon="🟢"
-          />
-          <StatCard
-            label="Stopped"
-            value={appStats.stopped}
-            icon="🟡"
-          />
-          <StatCard
-            label="Not Started"
-            value={appStats.notStarted}
-            icon="⚪"
-          />
+          {/* Hourly Activity Heatmap */}
+          {usage?.hourlyActivity && (
+            <HourlyActivityHeatmap hourlyActivity={usage.hourlyActivity} />
+          )}
         </div>
-      )}
-
-      {/* Hourly Activity Heatmap */}
-      {usage?.hourlyActivity && (
-        <HourlyActivityHeatmap hourlyActivity={usage.hourlyActivity} />
       )}
     </div>
   );
@@ -223,12 +214,12 @@ export default function Dashboard() {
 
 function StatCard({ label, value, icon }) {
   return (
-    <div className="bg-port-card border border-port-border rounded-lg p-3 sm:p-4" role="group" aria-label={`${label}: ${value}`}>
+    <div className="bg-port-bg border border-port-border rounded-lg p-3" role="group" aria-label={`${label}: ${value}`}>
       <div className="flex items-center gap-2 mb-1">
-        <span aria-hidden="true" className="text-base sm:text-lg">{icon}</span>
-        <span className="text-xs sm:text-sm text-gray-500">{label}</span>
+        <span aria-hidden="true" className="text-base">{icon}</span>
+        <span className="text-xs text-gray-500">{label}</span>
       </div>
-      <div className="text-xl sm:text-2xl font-bold text-white">{value}</div>
+      <div className="text-xl font-bold text-white">{value}</div>
     </div>
   );
 }
@@ -236,23 +227,19 @@ function StatCard({ label, value, icon }) {
 function HourlyActivityHeatmap({ hourlyActivity }) {
   const maxActivity = Math.max(...hourlyActivity, 1);
 
-  // Find peak hour(s)
   const peakValue = Math.max(...hourlyActivity);
   const peakHours = hourlyActivity
     .map((val, idx) => ({ hour: idx, count: val }))
     .filter(h => h.count === peakValue && h.count > 0);
 
-  // Calculate total sessions
   const totalSessions = hourlyActivity.reduce((sum, val) => sum + val, 0);
 
-  // Format hour label (12-hour format)
   const formatHour = (hour) => {
     if (hour === 0) return '12a';
     if (hour === 12) return '12p';
     return hour < 12 ? `${hour}a` : `${hour - 12}p`;
   };
 
-  // Get intensity class based on activity level
   const getIntensityClass = (count) => {
     if (count === 0) return 'bg-port-border/30';
     const intensity = count / maxActivity;
@@ -262,7 +249,6 @@ function HourlyActivityHeatmap({ hourlyActivity }) {
     return 'bg-port-success/20';
   };
 
-  // Determine peak time description
   const getPeakTimeDescription = () => {
     if (peakHours.length === 0 || peakValue === 0) return null;
     if (peakHours.length === 1) {
@@ -276,7 +262,7 @@ function HourlyActivityHeatmap({ hourlyActivity }) {
   if (totalSessions === 0) return null;
 
   return (
-    <div className="mt-8 bg-port-card border border-port-border rounded-xl p-4 sm:p-6">
+    <div className="bg-port-card border border-port-border rounded-xl p-4 sm:p-6">
       <div className="flex flex-col sm:flex-row sm:items-center justify-between gap-2 mb-4">
         <div className="flex items-center gap-3">
           <div className="text-2xl" aria-hidden="true">⏰</div>
@@ -292,7 +278,6 @@ function HourlyActivityHeatmap({ hourlyActivity }) {
         )}
       </div>
 
-      {/* Heatmap grid - 12 cols on sm+, 6 cols (2 rows) on mobile for larger touch targets */}
       <div className="grid grid-cols-6 sm:grid-cols-12 gap-1.5 sm:gap-1" role="img" aria-label="Hourly activity heatmap">
         {hourlyActivity.map((count, hour) => (
           <div
@@ -304,21 +289,15 @@ function HourlyActivityHeatmap({ hourlyActivity }) {
         ))}
       </div>
 
-      {/* Hour labels - simplified for mobile */}
       <div className="mt-2 grid grid-cols-6 sm:grid-cols-12 gap-1.5 sm:gap-1 text-xs text-gray-500">
         {hourlyActivity.map((_, hour) => (
-          <div
-            key={hour}
-            className="text-center"
-          >
-            {/* Mobile: show every 4th hour, Desktop: every 3rd hour */}
+          <div key={hour} className="text-center">
             <span className="hidden sm:inline">{hour % 3 === 0 ? formatHour(hour) : ''}</span>
             <span className="sm:hidden">{hour % 4 === 0 ? formatHour(hour) : ''}</span>
           </div>
         ))}
       </div>
 
-      {/* Legend */}
       <div className="mt-4 flex items-center justify-end gap-2 text-xs text-gray-500">
         <span>Less</span>
         <div className="flex gap-1">
