+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "portos",
-  "version": "0.25.7",
+  "version": "0.26.0",
   "private": true,
   "description": "Local dev machine App OS portal",
   "author": "Adam Eivy (@antic|@atomantic)",
